+++ L4D2AddonAssistant.Core/AddonNode.cs
@@ -98,13 +98,13 @@ namespace L4D2AddonAssistant
             }
         }
 
-        public IAddonNodeContainer Parent => ((IAddonNodeContainer?)Group) ?? Root; 
+        public IAddonNodeContainer Parent => ((IAddonNodeContainer?)Group) ?? Root;
 
         public AddonRoot Root => _root;
 
         public virtual bool RequireFile => false;
 
-        public bool IsAutoCheck => Root.IsAutoCheck; 
+        public bool IsAutoCheck => Root.IsAutoCheck;
 
         bool IHierarchyNode<AddonNode>.IsNonterminal => HasChildren;
 
@@ -180,13 +180,40 @@ namespace L4D2AddonAssistant
             }
         }
 
-        public string FullName => BuildFilePath(Group, Name);
+        public string FullName 
+        {
+            get 
+            {
+                if (Name.Length == 0)
+                {
+                    throw new InvalidOperationException("empty name");
+                }
+                return BuildFilePath(Group, Name); 
+            }
+        }
 
-        public string FileName => Name + FileExtension;
+        public string FileName 
+        {
+            get 
+            {
+                if (Name.Length == 0)
+                {
+                    throw new InvalidOperationException("empty name");
+                }
+                return Name + FileExtension; 
+            } 
+        }
 
         public string FilePath
         {
-            get => BuildFilePath(Group, FileName);
+            get 
+            {
+                if (Name.Length == 0)
+                {
+                    throw new InvalidOperationException("empty name");
+                }
+                return BuildFilePath(Group, FileName); 
+            }
         }
 
         public string FullFilePath => GetFullFilePath(FilePath);
@@ -333,6 +360,10 @@ namespace L4D2AddonAssistant
 
         public Task DestroyAsync()
         {
+            if (!IsValid)
+            {
+                return Task.CompletedTask;
+            }
             var tasks = new List<Task>();
             foreach (var node in this.GetSelfAndDescendantsByDfsPreorder())
             {