+++ L4D2AddonAssistant.GUI/IAppWindowManager.cs
@@ -15,5 +15,7 @@ namespace L4D2AddonAssistant
         void OpenDownloadListWindow();
 
         void OpenAboutWindow();
+
+        void OpenNewWorkshopCollectionWindow(AddonRoot addonRoot, AddonGroup? addonGroup);
     }
 }